refactor: extract hardcoded color constants to shared theme utilities
Consolidates theme color management by extracting hardcoded violet/purple
color values from 6 files into a single centralized theme utility module.

## Changes

**Created:**
- `src/components/shared/theme.ts` - Centralized theme utilities
  - `getCurrentThemeColors()`: Returns fg/bg/accent colors based on dark mode
  - `getVisualizationPalette(isDark)`: Returns CA visualization color palette
  - Full JSDoc documentation with type safety

**Refactored:**
- `src/components/shared/statsChart.ts` - Use `accentColor` from theme
- `src/components/desktop/layout.ts` - Replace `getCurrentColors()` with `getCurrentThemeColors()`
- `src/components/mobile/layout.ts` - Replace hardcoded palettes with `getVisualizationPalette()`
- `src/components/desktop/statistics.ts` - Use `accentColor` for chart colors

**Files NOT modified:**
- `src/components/desktop/tabContainer.ts` - No hardcoded colors found
- `src/components/desktop/patternInspector.ts` - No hardcoded colors found

## Impact

- **Duplication eliminated**: ~70 lines of hardcoded colors removed
- **Single source of truth**: All theme colors now managed in one place
- **Type safety**: TypeScript interface ensures consistent usage
- **Maintainability**: Theme changes only need updates in one file
- **Consistency**: All components use same color selection logic

## Verification

✅ No hardcoded violet colors remain outside theme.ts
✅ Linting passes (Biome auto-fix applied)
⚠️  Type check fails with pre-existing error in dataMode.ts:24 (not introduced by this PR)

## Notes

The TypeScript error in dataMode.ts existed before this refactoring - it's a
pre-existing issue where `createHeader()` expects arguments but is called
without them. This is unrelated to the theme refactoring.
